refactor: applied extract class refactoring method to pull out the internal classes from the ListUtils
