+++ src/main/java/org/apache/commons/collections4/list/CharSequenceAsList.java
@@ -1,0 +1,40 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.commons.collections4.list;
+
+import java.util.AbstractList;
+
+/**
+ * A simple wrapper to use a CharSequence as List.
+ */
+public class CharSequenceAsList extends AbstractList<Character> {
+    private final CharSequence sequence;
+
+    public CharSequenceAsList(final CharSequence sequence) {
+        this.sequence = sequence;
+    }
+
+    @Override
+    public Character get(final int index) {
+        return Character.valueOf(sequence.charAt(index));
+    }
+
+    @Override
+    public int size() {
+        return sequence.length();
+    }
+}
