+++ src/main/java/org/apache/commons/collections4/list/LcsVisitor.java
@@ -1,0 +1,52 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.commons.collections4.list;
+
+import org.apache.commons.collections4.sequence.CommandVisitor;
+
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * A helper class used to construct the longest common subsequence.
+ */
+public final class LcsVisitor<E> implements CommandVisitor<E> {
+    private final ArrayList<E> sequence;
+
+    public LcsVisitor() {
+        sequence = new ArrayList<>();
+    }
+
+    public List<E> getSubSequence() {
+        return sequence;
+    }
+
+    @Override
+    public void visitDeleteCommand(final E object) {
+        // noop
+    }
+
+    @Override
+    public void visitInsertCommand(final E object) {
+        // noop
+    }
+
+    @Override
+    public void visitKeepCommand(final E object) {
+        sequence.add(object);
+    }
+}
